QPID-482 : Small performance tweaks
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/branches/M2@541247 13f79535-47bb-0310-9956-ffa450edef68

+++ java/client/src/main/java/org/apache/qpid/client/AMQDestination.java
@@ -60,9 +60,9 @@ public abstract class AMQDestination implements Destination, Referenceable
     private static final int IS_EXCLUSIVE_MASK = 0x2;
     private static final int IS_AUTODELETE_MASK = 0x4;
 
-    public static final byte QUEUE_TYPE = 1;
-    public static final byte TOPIC_TYPE = 2;
-    public static final byte UNKNOWN_TYPE = 3;
+    public static final Integer QUEUE_TYPE = Integer.valueOf(1);
+    public static final Integer TOPIC_TYPE = Integer.valueOf(2);
+    public static final Integer UNKNOWN_TYPE = Integer.valueOf(3);
 
     protected AMQDestination(String url) throws URLSyntaxException
     {
@@ -213,7 +213,7 @@ public abstract class AMQDestination implements Destination, Referenceable
     }
 
     public String toURL()
-    {
+    {        
         String url = _url;
         if(url == null)
         {
